PYTHON-6018 Consolidate backpressure helpers into helpers_shared.py
