+++ test/test_retryable_writes.py
@@ -831,10 +831,7 @@ class TestErrorPropagationAfterEncounteringMultipleErrors(IntegrationTest):
         self.assertEqual(len(started_inserts), MAX_ADAPTIVE_RETRIES + 1)
 
     def test_backoff_is_not_applied_for_non_overload_errors(self):
-        if _IS_SYNC:
-            mock_target = "pymongo.synchronous.helpers._RetryPolicy.backoff"
-        else:
-            mock_target = "pymongo.helpers._RetryPolicy.backoff"
+        mock_target = "pymongo.helpers_shared._RetryPolicy.backoff"
 
         # Create a client.
         listener = OvertCommandListener()
